requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cssmin==0.2.0
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
email-validator==2.3.0
exceptiongroup==1.3.0
faiss-cpu==1.12.0
filelock==3.13.1
Flask==3.1.2
Flask-Assets==2.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Scss==0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fsspec==2024.6.1
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
libsass==0.23.0
Mako==1.3.10
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.3
numpy==2.2.6
openai==1.102.0
packaging==25.0
pillow==11.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyScss==1.4.0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
sentence-transformers==5.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sympy==1.13.3
threadpoolctl==3.6.0
tokenizers==0.22.0
tomli==2.2.1
torch==2.8.0+cpu
torchaudio==2.8.0+cpu
torchvision==0.23.0+cpu
tqdm==4.67.1
transformers==4.56.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
webassets==3.0.0
Werkzeug==3.1.3
WTForms==3.2.1
zope.dottedname==6.1
